Add multiple sessions
App.js now just adds session configured in session.config.js
Added express-session to deal with the problem with long cookies
At the moment selecting the session is done through a code change I just
can't think of a better place for it right now
